# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Hexpertify Therapist Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clients:
    get:
      operationId: getClients
      tags: [clients]
      summary: Get client list
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, completed, inactive, high_priority, new]
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Client list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get client workspace detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientDetail"
  /clients/{id}/assessments:
    get:
      operationId: getClientAssessments
      tags: [clients]
      summary: Get client assessments
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client assessments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Assessment"
  /clients/{id}/mood:
    get:
      operationId: getClientMood
      tags: [clients]
      summary: Get client mood tracking data
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Mood tracking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MoodTracking"
  /clients/{id}/homework:
    get:
      operationId: getClientHomework
      tags: [clients]
      summary: Get client homework and activities
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Homework list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HomeworkItem"
  /clients/{id}/session-history:
    get:
      operationId: getClientSessionHistory
      tags: [clients]
      summary: Get client session history
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Session history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SessionHistory"
...
  /outcomes/individual/{clientId}:
    get:
      operationId: fetchClientOutcomes
      tags: [outcomes]
      summary: Get individual client outcomes
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client outcomes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientOutcomes"
  /outcomes/caseload:
    get:
      operationId: getCaseloadOutcomes
      tags: [outcomes]
      summary: Get therapist caseload outcomes summary
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [week, month, year, custom]
      responses:
        "200":
          description: Caseload outcomes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseloadOutcomes"
...
  /blog/posts:
    post:
      operationId: submitBlogPost
      tags: [blog]
      summary: Submit a full blog post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlogPostInput"
      responses:
        "201":
          description: Blog post submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
    get:
      operationId: getBlogPosts
      tags: [blog]
      summary: Get all blog posts
      responses:
        "200":
          description: Blog posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BlogPost"
  /blog/outlines:
    post:
      operationId: submitBlogOutline
      tags: [blog]
      summary: Submit a blog outline
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlogOutlineInput"
      responses:
        "201":
          description: Blog outline submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogOutline"
    get:
      operationId: getBlogOutlines
      tags: [blog]
      summary: Get all blog outlines
      responses:
        "200":
          description: Blog outlines
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BlogOutline"
...
components:
  schemas:
...
    Client:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        initials:
          type: string
        age:
          type: integer
        gender:
          type: string
        status:
          type: string
          enum: [active, completed, inactive, high_priority, new]
        primaryGoal:
          type: string
        lastSession:
          type: ["string", "null"]
        nextSession:
          type: ["string", "null"]
        progressPercent:
          type: integer
      required:
        - id
        - name
        - initials
        - age
        - gender
        - status
        - primaryGoal
        - progressPercent
    ClientDetail:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        initials:
          type: string
        age:
          type: integer
        gender:
          type: string
        status:
          type: string
        primaryGoal:
          type: string
        presentingProblems:
          type: array
          items:
            type: string
        identifiedConcerns:
          type: array
          items:
            type: string
        therapyGoals:
          type: array
          items:
            type: string
        preferredLanguage:
          type: string
        communicationPreference:
          type: string
        therapyTimeline:
          type: string
        aiIntakeSummary:
          type: string
        sessionCount:
          type: integer
        startDate:
          type: string
      required:
        - id
        - name
        - initials
        - age
        - gender
        - status
        - primaryGoal
        - presentingProblems
        - identifiedConcerns
        - therapyGoals
        - preferredLanguage
        - communicationPreference
        - therapyTimeline
        - aiIntakeSummary
        - sessionCount
        - startDate
    Assessment:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [PHQ-9, GAD-7, PANAS, other]
        name:
          type: string
        currentScore:
          type: number
        maxScore:
          type: number
        previousScore:
          type: ["number", "null"]
        severity:
          type: string
        completedAt:
          type: string
        trend:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
              score:
                type: number
            required: [date, score]
      required:
        - id
        - type
        - name
        - currentScore
        - maxScore
        - severity
        - completedAt
        - trend
    MoodTracking:
      type: object
      properties:
        today:
          type: ["number", "null"]
        weeklyTrend:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
              mood:
                type: number
              note:
                type: ["string", "null"]
            required: [date, mood]
      required: [weeklyTrend]
    HomeworkItem:
      type: object
      properties:
        id:
          type: integer
        activity:
          type: string
        instructions:
          type: string
        frequency:
          type: string
        dueDate:
          type: string
        status:
          type: string
          enum: [completed, pending, missed]
        completionPercent:
          type: integer
        streak:
          type: integer
        clientReflection:
          type: ["string", "null"]
      required:
        - id
        - activity
        - instructions
        - frequency
        - dueDate
        - status
        - completionPercent
        - streak
    SessionHistory:
      type: object
      properties:
        id:
          type: integer
        date:
          type: string
        durationMinutes:
          type: integer
        sessionType:
          type: string
        summary:
          type: string
        homeworkAssigned:
          type: ["string", "null"]
        therapistNotes:
          type: ["string", "null"]
      required:
        - id
        - date
        - durationMinutes
        - sessionType
        - summary
...
    ClientOutcomes:
      type: object
      properties:
        clientId:
          type: integer
        clientName:
          type: string
        improvementScore:
          type: number
        goalAchievementRate:
          type: number
        totalGoals:
          type: integer
        completedGoals:
          type: integer
        goalsInProgress:
          type: integer
        attendancePercent:
          type: number
        sessionsAttended:
          type: integer
        missedSessions:
          type: integer
        rescheduledSessions:
          type: integer
        homeworkCompletionPercent:
          type: number
        assignedActivities:
          type: integer
        completedActivities:
          type: integer
        currentStreak:
          type: integer
        engagementScore:
          type: number
        engagementLevel:
          type: string
          enum: [high, medium, low]
        assessmentTrends:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              data:
                type: array
                items:
                  type: object
                  properties:
                    date:
                      type: string
                    score:
                      type: number
                  required: [date, score]
            required: [name, data]
      required:
        - clientId
        - clientName
        - improvementScore
        - goalAchievementRate
        - totalGoals
        - completedGoals
        - goalsInProgress
        - attendancePercent
        - sessionsAttended
        - missedSessions
        - rescheduledSessions
        - homeworkCompletionPercent
        - assignedActivities
        - completedActivities
        - currentStreak
        - engagementScore
        - engagementLevel
        - assessmentTrends
    CaseloadOutcomes:
      type: object
      properties:
        averageImprovementScore:
          type: number
        averageGoalAchievementRate:
          type: number
        averageAssessmentImprovement:
          type: number
        overallAttendanceRate:
          type: number
        averageHomeworkAdherence:
          type: number
        averageEngagementScore:
          type: number
        period:
          type: string
        clientBreakdown:
          type: array
          items:
            type: object
            properties:
              clientId:
                type: integer
              clientName:
                type: string
              improvementScore:
                type: number
              engagementScore:
                type: number
              status:
                type: string
            required: [clientId, clientName, improvementScore, engagementScore, status]
      required:
        - averageImprovementScore
        - averageGoalAchievementRate
        - averageAssessmentImprovement
        - overallAttendanceRate
        - averageHomeworkAdherence
        - averageEngagementScore
        - period
        - clientBreakdown
...
    BlogPostInput:
      type: object
      properties:
        title:
          type: string
        featuredImage:
          type: ["string", "null"]
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        content:
          type: string
      required: [title, category, content]
    BlogPost:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        content:
          type: string
        status:
          type: string
          enum: [draft, submitted, published]
        createdAt:
          type: string
      required: [id, title, category, tags, content, status, createdAt]
    BlogOutlineInput:
      type: object
      properties:
        proposedTitle:
          type: string
        keyPoints:
          type: array
          items:
            type: string
        targetAudience:
          type: string
        keywords:
          type: array
          items:
            type: string
        notes:
          type: ["string", "null"]
      required: [proposedTitle, keyPoints, targetAudience, keywords]
    BlogOutline:
      type: object
      properties:
        id:
          type: integer
        proposedTitle:
          type: string
        keyPoints:
          type: array
          items:
            type: string
        targetAudience:
          type: string
        keywords:
          type: array
          items:
            type: string
        notes:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, approved, rejected]
        createdAt:
          type: string
      required:
        - id
        - proposedTitle
        - keyPoints
        - targetAudience
        - keywords
        - status
        - createdAt
...