# lib/api-zod/src/generated/types/homeworkItem.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Hexpertify Therapist Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { HomeworkItemStatus } from './homeworkItemStatus';

export interface HomeworkItem {
  id: number;
  activity: string;
  instructions: string;
  frequency: string;
  dueDate: string;
  status: HomeworkItemStatus;
  completionPercent: number;
  streak: number;
  /** @nullable */
  clientReflection?: string | null;
}